Moves generate ballot to package structure
