+++ src/packages/generate-ballot/components/sections/candidate.tsx
