+++ src/packages/generate-ballot/components/sections/party-header.tsx
