+++ src/pages/generate-ballot.tsx
@@ -1,131 +1,18 @@
 import React, { useEffect } from "react"
 
-import { Machine, assign, EventObject } from "xstate"
+import { EventObject } from "xstate"
 import { useMachine } from "@xstate/react"
 
 import {
   GovernmentalBallot,
   LegislativeBallot,
   MunicipalBallot,
-} from "../components/ballots/index"
-import { PUBLIC_S3_BUCKET } from "../packages/practica/services/constants"
+} from "../packages/generate-ballot/components/index"
 import Switch from "../components/switch"
 import Case from "../components/case"
 import Default from "../components/default"
-
-type BallotContent = {
-  ocrResult: string
-  logoImg?: string
-}
-
-type VotesCoordinates = {
-  column: number
-  row: number
-}
-
-type BallotMachineContext = {
-  type: string | null
-  path: string | null
-  votes: VotesCoordinates[]
-  ballot: BallotContent[][]
-}
-
-async function fetchBallot(path: string | null) {
-  if (path) {
-    const resp = await fetch(`${PUBLIC_S3_BUCKET}${path}/data.json`)
-    const data = await resp.json()
-
-    return data
-  }
-
-  throw Error("Invalid ballot path")
-}
-
-const BallotMachine = Machine<BallotMachineContext>({
-  id: "ballotMachine",
-  initial: "idle",
-  context: {
-    type: "",
-    path: "",
-    ballot: [],
-    votes: [],
-  },
-  states: {
-    idle: {
-      on: {
-        FETCH: "loading",
-      },
-    },
-    loading: {
-      invoke: {
-        id: "fetchBallot",
-        src: (context: BallotMachineContext) => fetchBallot(context.path),
-        onDone: {
-          target: "success",
-          actions: assign({ ballot: (_, event) => event.data }),
-        },
-        onError: {
-          target: "failure",
-        },
-      },
-    },
-    success: {
-      id: "generateBallotMachine",
-      initial: "idle",
-      states: {
-        idle: {
-          on: {
-            "": [
-              {
-                target: "governmental",
-                cond(context) {
-                  return context.type === "estatal"
-                },
-              },
-              {
-                target: "legislative",
-                cond(context) {
-                  return context.type === "legislativa"
-                },
-              },
-              {
-                target: "municipal",
-                cond(context) {
-                  return context.type === "municipal"
-                },
-              },
-              {
-                target: "unknown",
-                cond(context) {
-                  return (
-                    context.type !== "estatal" &&
-                    context.type !== "legislativa" &&
-                    context.type !== "municipal"
-                  )
-                },
-              },
-            ],
-          },
-        },
-        governmental: {
-          type: "final",
-        },
-        legislative: {
-          type: "final",
-        },
-        municipal: {
-          type: "final",
-        },
-        unknown: {
-          type: "final",
-        },
-      },
-    },
-    failure: {
-      type: "final",
-    },
-  },
-})
+import { BallotMachine } from "../packages/generate-ballot/machines/ballot-machine"
+import { BallotMachineContext } from "../packages/generate-ballot/types/ballot-machine"
 
 type PageProps = {
   location: Location
@@ -165,14 +52,16 @@ export default function GenerateBallot({ location }: PageProps) {
         </Case>
         <Case value={{ success: "legislative" }}>
           <LegislativeBallot
-            ballotPath={state.context.path}
-            votes={state.context.ballot}
+            path={state.context.path}
+            structure={state.context.ballot}
+            votes={votes}
           />
         </Case>
         <Case value={{ success: "municipal" }}>
           <MunicipalBallot
-            ballotPath={state.context.path}
-            votes={state.context.ballot}
+            path={state.context.path}
+            structure={state.context.ballot}
+            votes={votes}
           />
         </Case>
         <Default>
